chatbot2/pyproject.toml
[tool.poetry]
name = "pacote-de-pdf"
version = "0.1.0"
description = "subir_pdf"
authors = ["enio"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.37.1"
langchain = "^0.2.12"
python-dotenv = "^1.0.1"
pypdf2 = "^3.0.1"
faiss-cpu = "^1.8.0.post1"
langchain-community = "^0.2.11"
tiktoken = "^0.7.0"
openai = "0.28"
streamlit-chat = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
